refactor: centralize db connection and remove supabase setup
