+++ backend/api/config/__init__.py
@@ -1,7 +1,7 @@
 # api/config/__init__.py
 # Inizializzazione del package config
 
-from .database import get_supabase_client, db_config
+from .database import get_supabase_client, db_config, get_db_connection
 
-__all__ = ['get_supabase_client', 'db_config']
+__all__ = ['get_supabase_client', 'db_config', 'get_db_connection']
